requirements.txt
# Core ADK dependencies
google-adk>=0.1.0
google-generativeai>=0.3.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
slowapi>=0.1.9

# Data processing
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.10.0
pylint>=3.0.0
mypy>=1.6.0
isort>=5.12.0

# Logging and monitoring
structlog>=23.2.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
